Split TreeNode, TreeStyle and TreeBuilder into seperate files
